requirements.txt
fastapi
uvicorn
requests
python-dotenv
numpy
pandas
python-telegram-bot>=21
MetaTrader5 ; sys_platform == 'win32'
